pyproject.toml
[tool.poetry]
name = "ride-sharing-service"
version = "0.1.0"
description = "Сервис для организации совместных поездок"
authors = ["BootyOne <BootyOne@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ride_sharing_service"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
python-dotenv = "^1.0.0"
peewee = "^3.16.3"
alembic = "^1.11.3"
peewee-async = "^0.8.1"
uvicorn = "^0.23.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
asyncpg = "^0.28.0"
aiopg = "^1.4.0"
bcrypt = "^4.0.1"
peewee-migrate = "^1.12.2"
pydantic = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
